Apply a fix that was already in reverse_https
git-svn-id: file:///home/svn/framework3/trunk@13205 4d416f70-5f16-0410-b530-b9f4589650da

+++ lib/msf/core/handler/reverse_http.rb
@@ -105,7 +105,7 @@ module ReverseHttp
 	# active on sub-urls.
 	#
 	def stop_handler
-		self.service.remove_resource("/")
+		self.service.remove_resource("/") if self.service
 	end
 
 	attr_accessor :service # :nodoc:
